Experimental versioned route handlers
This is based off the Nova idea of a VersionedMethod but altered
to work with the flask style of declaring routes. On the one hand
this is rather fun. On the other it is completely insane.

One could ask why not just dispatch on version within the handler
instead of versioning the handlers.